+++ hrp/step_mobile_ui.go
@@ -601,9 +601,9 @@ func runStepMobileUI(s *SessionRunner, step *TStep) (stepResult *StepResult, err
 
 			// check if app is in foreground
 			packageName := uiDriver.Driver.GetLastLaunchedApp()
-			yes, err2 := uiDriver.Driver.IsAppInForeground(packageName)
-			if packageName != "" && (!yes || err2 != nil) {
-				log.Error().Err(err2).Str("packageName", packageName).Msg("app is not in foreground")
+			err := uiDriver.Driver.AssertAppForeground(packageName)
+			if packageName != "" && err != nil {
+				log.Error().Err(err).Str("packageName", packageName).Msg("app is not in foreground")
 				err = errors.Wrap(code.MobileUIAppNotInForegroundError, err.Error())
 			}
 		}